add test case and fix bug

+++ include/typingtest.h
@@ -1,6 +1,11 @@
 #ifndef TYPINGTEST_H
 #define TYPINGTEST_H
 
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <time.h>
+
 #define ANSI_CLEAR_LINE "\033[2K"
 #define ANSI_RESET_CURSOR "\033[1G"
 #define ANSI_COLOR_GREEN "\033[32m"
@@ -15,6 +20,6 @@ typedef struct {
 void loadParagraphs(TypingTest* typingTest, const char* filename);
 void runTypingTest(TypingTest* typingTest);
 char* getRandomParagraph(TypingTest* typingTest);
-void printParagraph(const char* paragraph, int currentPosition);
+void printParagraph(const char* paragraph, size_t currentPosition);
 
 #endif
